requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
pydantic
requests
python-dotenv
pytest
